+++ src/main/java/ar/edu/unlam/tallerweb1/delivery/DatosMascotas.java
@@ -12,9 +12,9 @@ public class DatosMascotas {
     private String imagen;
 
     private Long tipo;
-    private Long idUsuario;
+    private int idUsuario;
 
-    public DatosMascotas(String nombre, String descripcion, String imagen, String raza, Long estado, Long tipo, Long idUsuario){
+    public DatosMascotas(String nombre, String descripcion, String imagen, String raza, Long estado, Long tipo, int idUsuario){
         this.nombre = nombre;
         this.descripcion = descripcion;
         this.imagen = imagen;
@@ -84,11 +84,11 @@ public class DatosMascotas {
     }
 
 
-    public Long getIdUsuario() {
+    public int getIdUsuario() {
         return idUsuario;
     }
 
-    public void setIdUsuario(Long idUsuario) {
+    public void setIdUsuario(int idUsuario) {
         this.idUsuario = idUsuario;
     }
